enable adding categorical data array to layer's underlying data; reorganize construction of categorical colormaps to facilitate this
