+++ drilldown/utils.py
@@ -1,6 +1,8 @@
 import collections.abc
 import numpy as np
 import pandas as pd
+import distinctipy
+from matplotlib.colors import ListedColormap
 
 
 def convert_to_numpy_array(arr, collapse_dim=True):
@@ -23,6 +25,98 @@ def convert_to_numpy_array(arr, collapse_dim=True):
         raise TypeError("Data must be a sequence or Pandas object.")
 
     return arr
+
+
+def convert_array_type(arr, return_type=False):
+    try:
+        arr = arr.astype("float")
+        _type = "float"
+    except ValueError:
+        arr = arr.astype("str")
+        _type = "str"
+
+    if return_type == True:
+        return arr, _type
+    else:
+        return arr
+
+
+def encode_categorical_data(data):
+    data = convert_to_numpy_array(data)
+
+    data_encoded, categories = pd.factorize(data)
+    codes = np.arange(len(categories))
+
+    # convert codes to float
+    codes = np.array(codes, dtype="float")
+    data_encoded = np.array(data_encoded, dtype="float")
+
+    # center numerical representation of categorical data, while maintaining range, to address pyvista's color mapping querks
+    codes[1:-1] += 0.5
+    data_encoded[
+        (data_encoded != data_encoded.min()) & (data_encoded != data_encoded.max())
+    ] += 0.5
+    code_to_cat_map = {code: cat for code, cat in zip(codes, categories)}
+
+    return code_to_cat_map, data_encoded
+
+
+def make_matplotlib_categorical_color_map(colors):
+    mapping = np.linspace(0, len(colors) - 1, 256)
+    new_colors = np.empty((256, 3))
+    i_pre = -np.inf
+    for i, color in enumerate(colors):
+        new_colors[(mapping > i_pre) & (mapping <= i)] = list(color)
+        i_pre = i
+    map = ListedColormap(colors[0:-1])
+    map.set_under(colors[0])
+    map.set_over(colors[-1])
+    return map
+
+
+def get_cycled_colors(n):
+    from itertools import cycle
+    import matplotlib as mpl
+
+    cycler = mpl.rcParams["axes.prop_cycle"]
+    colors = cycle(cycler)
+    colors = [next(colors)["color"] for i in range(n)]
+    colors = [mpl.colors.to_rgb(color) for color in colors]
+
+    return colors
+
+
+def make_color_map_fractional(map):
+    for name in map.keys():
+        if max(map[name]) > 1:
+            map[name] = tuple([val / 255 for val in map[name]])
+        elif max(map[name]) >= 0:
+            # raise ValueError("Color is already fractional.")
+            pass
+        else:
+            raise ValueError("Color values cannot be negative.")
+
+    return map
+
+
+def make_categorical_cmap(categories, cycle=True, rng=999, pastel_factor=0.2):
+    n_colors = len(categories)
+    if cycle == True:
+        colors = get_cycled_colors(n_colors)
+
+    elif cycle == False:
+        colors = distinctipy.get_colors(
+            n_colors,
+            pastel_factor=pastel_factor,
+            rng=rng,
+        )
+
+    # create categorical color map
+    cat_to_color_map = {cat: color for cat, color in zip(categories, colors)}
+    # create matplotlib categorical color map
+    matplotlib_formatted_color_maps = make_matplotlib_categorical_color_map(colors)
+
+    return cat_to_color_map, matplotlib_formatted_color_maps
 
 
 def is_jupyter():
